Breaking error types out of JDP database source.
