+++ databases/jdp/database.go
@@ -44,289 +44,6 @@ import (
 	"github.com/kbase/dts/frictionless"
 )
 
-const (
-	jdpBaseURL     = "https://files.jgi.doe.gov/"
-	filePathPrefix = "/global/dna/dm_archive/" // directory containing JDP files
-)
-
-// this error type is returned when a file is requested for which the requester
-// does not have permission
-type PermissionDeniedError struct {
-	fileId string
-}
-
-func (e PermissionDeniedError) Error() string {
-	return fmt.Sprintf("Can't access file %s: permission denied.", e.fileId)
-}
-
-// this error type is returned when a file is requested and is not found
-type FileIdNotFoundError struct {
-	fileId string
-}
-
-func (e FileIdNotFoundError) Error() string {
-	return fmt.Sprintf("Can't access file %s: not found.", e.fileId)
-}
-
-// a mapping from file suffixes to format labels
-var suffixToFormat = map[string]string{
-	"bam":      "bam",
-	"bam.bai":  "bai",
-	"blasttab": "blast",
-	"bz":       "bzip",
-	"bz2":      "bzip2",
-	"csv":      "csv",
-	"faa":      "fasta",
-	"fasta":    "fasta",
-	"fasta.gz": "fasta",
-	"fastq":    "fastq",
-	"fastq.gz": "fastq",
-	"fna":      "fasta",
-	"gff":      "gff",
-	"gff3":     "gff3",
-	"gz":       "gz",
-	"html":     "html",
-	"info":     "texinfo",
-	"out":      "text",
-	"pdf":      "pdf",
-	"tar":      "tar",
-	"tar.gz":   "tar",
-	"tar.bz":   "tar",
-	"tar.bz2":  "tar",
-	"tsv":      "tsv",
-	"txt":      "text",
-}
-
-// this gets populated automatically with the keys in suffixToFormat
-var supportedSuffixes []string
-
-// a mapping from file format labels to mime types
-var formatToMimeType = map[string]string{
-	"bam":   "application/octet-stream",
-	"bai":   "application/octet-stream",
-	"csv":   "text/csv",
-	"fasta": "text/plain",
-	"fastq": "text/plain",
-	"gff":   "text/plain",
-	"gff3":  "text/plain",
-	"gz":    "application/gzip",
-	"bz":    "application/x-bzip",
-	"bz2":   "application/x-bzip2",
-	"tar":   "application/x-tar",
-	"text":  "text/plain",
-}
-
-// a mapping from the JDP's reported file types to mime types
-// (backup method for determining mime types)
-var fileTypeToMimeType = map[string]string{
-	"text":     "text/plain",
-	"fasta":    "text/plain",
-	"fasta.gz": "application/gzip",
-	"fastq":    "text/plain",
-	"fastq.gz": "application/gzip",
-	"tab":      "text/plain",
-	"tar.gz":   "application/x-tar",
-	"tar.bz":   "application/x-tar",
-	"tar.bz2":  "application/x-tar",
-}
-
-// extracts the file format from the name and type of the file
-func formatFromFileName(fileName string) string {
-	// make a list of the supported suffixes if we haven't yet
-	if supportedSuffixes == nil {
-		supportedSuffixes = make([]string, 0)
-		for suffix := range suffixToFormat {
-			supportedSuffixes = append(supportedSuffixes, suffix)
-		}
-	}
-
-	// determine whether the file matches any of the supported suffixes,
-	// selecting the longest matching suffix
-	format := "unknown"
-	longestSuffix := 0
-	for _, suffix := range supportedSuffixes {
-		if strings.HasSuffix(fileName, suffix) && len(suffix) > longestSuffix {
-			format = suffixToFormat[suffix]
-			longestSuffix = len(suffix)
-		}
-	}
-	return format
-}
-
-// extracts the file format from the name and type of the file
-func mimeTypeFromFormatAndTypes(format string, fileTypes []string) string {
-	// try to match the file type to a mime type
-	for _, fileType := range fileTypes {
-		if mimeType, ok := fileTypeToMimeType[fileType]; ok {
-			return mimeType
-		}
-	}
-	// check the file format to see whether it matches a mime type
-	if mimeType, ok := formatToMimeType[format]; ok {
-		return mimeType
-	}
-	return ""
-}
-
-// extracts file type information from the given File
-func fileTypesFromFile(file File) []string {
-	// TODO: See https://pkg.go.dev/encoding/json?utm_source=godoc#example-RawMessage-Unmarshal
-	// TODO: for an example of how to unmarshal a variant type.
-	return []string{}
-}
-
-// extracts source information from the given metadata
-func sourcesFromMetadata(md Metadata) []frictionless.DataSource {
-	sources := make([]frictionless.DataSource, 0)
-	piInfo := md.Proposal.PI
-	if len(piInfo.LastName) > 0 {
-		var title string
-		if len(piInfo.FirstName) > 0 {
-			title = fmt.Sprintf("%s, %s", piInfo.LastName, piInfo.FirstName)
-			if len(piInfo.MiddleName) > 0 {
-				title += fmt.Sprintf(" %s", piInfo.MiddleName)
-			}
-			if len(piInfo.Institution) > 0 {
-				if len(piInfo.Country) > 0 {
-					title += fmt.Sprintf(" (%s, %s)", piInfo.Institution, piInfo.Country)
-				} else {
-					title += fmt.Sprintf(" (%s)", piInfo.Institution)
-				}
-			} else if len(piInfo.Country) > 0 {
-				title += fmt.Sprintf(" (%s)", piInfo.Country)
-			}
-		}
-		var doiURL string
-		if len(md.Proposal.AwardDOI) > 0 {
-			doiURL = fmt.Sprintf("https://doi.org/%s", md.Proposal.AwardDOI)
-		}
-		source := frictionless.DataSource{
-			Title: title,
-			Path:  doiURL,
-			Email: piInfo.EmailAddress,
-		}
-		sources = append(sources, source)
-	}
-	return sources
-}
-
-// creates a Frictionless DataResource-savvy name for a file:
-// * the name consists of lower case characters plus '.', '-', and '_'
-// * all forbidden characters encountered in the filename are removed
-// * a number suffix is added if needed to make the name unique
-func dataResourceName(filename string) string {
-	name := strings.ToLower(filename)
-
-	// remove any file suffix
-	lastDot := strings.LastIndex(name, ".")
-	if lastDot != -1 {
-		name = name[:lastDot]
-	}
-
-	// replace sequences of invalid characters with '_'
-	for {
-		isInvalid := func(c rune) bool {
-			return !unicode.IsLetter(c) && !unicode.IsDigit(c) && c != '_' && c != '-' && c != '.'
-		}
-		start := strings.IndexFunc(name, isInvalid)
-		if start >= 0 {
-			nameRunes := []rune(name)
-			end := start + 1
-			for end < len(name) && isInvalid(nameRunes[end]) {
-				end++
-			}
-			if end < len(name) {
-				name = name[:start] + string('_') + name[end+1:]
-			} else {
-				name = name[:start] + string('_')
-			}
-		} else {
-			break
-		}
-	}
-
-	return name
-}
-
-// creates a DataResource from a File
-func dataResourceFromFile(file File) frictionless.DataResource {
-	id := "JDP:" + file.Id
-	format := formatFromFileName(file.Name)
-	fileTypes := fileTypesFromFile(file)
-	sources := sourcesFromMetadata(file.Metadata)
-
-	// we use relative file paths in accordance with the Frictionless
-	// Data Resource specification
-	filePath := filepath.Join(strings.TrimPrefix(file.Path, filePathPrefix), file.Name)
-
-	pi := file.Metadata.Proposal.PI
-	return frictionless.DataResource{
-		Id:        id,
-		Name:      dataResourceName(file.Name),
-		Path:      filePath,
-		Format:    format,
-		MediaType: mimeTypeFromFormatAndTypes(format, fileTypes),
-		Bytes:     file.Size,
-		Hash:      file.MD5Sum,
-		Sources:   sources,
-		Credit: credit.CreditMetadata{
-			Identifier:   id,
-			ResourceType: "dataset",
-			Titles: []credit.Title{
-				{
-					Title: filePath,
-				},
-			},
-			Dates: []credit.EventDate{
-				{
-					Date:  file.Date,
-					Event: "Created",
-				},
-				{
-					Date:  file.AddedDate,
-					Event: "Accepted",
-				},
-				{
-					Date:  file.ModifiedDate,
-					Event: "Updated",
-				},
-			},
-			Publisher: credit.Organization{
-				OrganizationId:   "ROR:04xm1d337",
-				OrganizationName: "Joint Genome Institute",
-			},
-			RelatedIdentifiers: []credit.PermanentID{
-				{
-					Id:               file.Metadata.Proposal.DOI,
-					Description:      "Proposal DOI",
-					RelationshipType: "IsCitedBy",
-				},
-				{
-					Id:               file.Metadata.Proposal.AwardDOI,
-					Description:      "Awarded proposal DOI",
-					RelationshipType: "IsCitedBy",
-				},
-			},
-			Contributors: []credit.Contributor{
-				{
-					ContributorType: "Person",
-					// ContributorId: nothing yet
-					Name:       strings.TrimSpace(fmt.Sprintf("%s, %s %s", pi.LastName, pi.FirstName, pi.MiddleName)),
-					GivenName:  strings.TrimSpace(fmt.Sprintf("%s %s", pi.FirstName, pi.MiddleName)),
-					FamilyName: strings.TrimSpace(pi.LastName),
-					Affiliations: []credit.Organization{
-						{
-							OrganizationName: pi.Institution,
-						},
-					},
-					ContributorRoles: "PI",
-				},
-			},
-			Version: file.Date,
-		},
-	}
-}
-
 // file database appropriate for handling JDP searches and transfers
 // (implements the databases.Database interface)
 type Database struct {
@@ -375,144 +92,6 @@ func NewDatabase(orcid string) (databases.Database, error) {
 	}, nil
 }
 
-// adds an appropriate authorization header to given HTTP request
-func (db Database) addAuthHeader(request *http.Request) {
-	if len(db.Secret) > 0 { // use shared secret
-		request.Header.Add("Authorization", fmt.Sprintf("Token %s_%s", db.Orcid, db.Secret))
-	} else { // try SSO token
-		request.Header.Add("Authorization", fmt.Sprintf("Bearer %s", db.SsoToken))
-	}
-}
-
-// performs a GET request on the given resource, returning the resulting
-// response and error
-func (db *Database) get(resource string, values url.Values) (*http.Response, error) {
-	var u *url.URL
-	u, err := url.ParseRequestURI(jdpBaseURL)
-	if err != nil {
-		return nil, err
-	}
-	u.Path = resource
-	u.RawQuery = values.Encode()
-	res := fmt.Sprintf("%v", u)
-	slog.Debug(fmt.Sprintf("GET: %s", res))
-	req, err := http.NewRequest(http.MethodGet, res, http.NoBody)
-	if err != nil {
-		return nil, err
-	}
-	db.addAuthHeader(req)
-	return db.Client.Do(req)
-}
-
-// performs a POST request on the given resource, returning the resulting
-// response and error
-func (db *Database) post(resource string, body io.Reader) (*http.Response, error) {
-	u, err := url.ParseRequestURI(jdpBaseURL)
-	if err != nil {
-		return nil, err
-	}
-	u.Path = resource
-	res := fmt.Sprintf("%v", u)
-	slog.Debug(fmt.Sprintf("POST: %s", res))
-	req, err := http.NewRequest(http.MethodPost, res, body)
-	if err != nil {
-		return nil, err
-	}
-	db.addAuthHeader(req)
-	req.Header.Set("Content-Type", "application/json")
-	return db.Client.Do(req)
-}
-
-// this helper extracts files for the JDP /search GET query with given parameters
-func (db *Database) filesFromSearch(params url.Values) (databases.SearchResults, error) {
-	var results databases.SearchResults
-
-	idEncountered := make(map[string]bool) // keep track of duplicates
-
-	// extract any requested "extra" metadata fields (and scrub them from params)
-	var extraFields []string
-	if params.Has("extra") {
-		extraFields = strings.Split(params.Get("extra"), ",")
-		params.Del("extra")
-	}
-
-	resp, err := db.get("search", params)
-	if err != nil {
-		return results, err
-	}
-	defer resp.Body.Close()
-	var body []byte
-	body, err = io.ReadAll(resp.Body)
-	if err != nil {
-		return results, err
-	}
-	type JDPResults struct {
-		Organisms []struct {
-			Id    string `json:"id"`
-			Files []File `json:"files"`
-		} `json:"organisms"`
-	}
-	results.Resources = make([]frictionless.DataResource, 0)
-	var jdpResults JDPResults
-	err = json.Unmarshal(body, &jdpResults)
-	if err != nil {
-		return results, err
-	}
-	for _, org := range jdpResults.Organisms {
-		resources := make([]frictionless.DataResource, 0)
-		for _, file := range org.Files {
-			res := dataResourceFromFile(file)
-
-			// add any requested additional metadata
-			if extraFields != nil {
-				extras := "{"
-				for i, field := range extraFields {
-					if i > 0 {
-						extras += ", "
-					}
-					switch field {
-					case "project_id":
-						extras += fmt.Sprintf(`"project_id": "%s"`, org.Id)
-					case "img_taxon_oid":
-						extras += fmt.Sprintf(`"img_taxon_oid": %d`, file.Metadata.IMG.TaxonOID)
-					}
-				}
-				extras += "}"
-				res.Extra = json.RawMessage(extras)
-			}
-
-			// add the resource to our results if it's not there already
-			if _, encountered := idEncountered[res.Id]; !encountered {
-				resources = append(resources, res)
-				idEncountered[res.Id] = true
-			}
-		}
-		results.Resources = append(results.Resources, resources...)
-	}
-	return results, nil
-}
-
-// returns the page number and page size corresponding to the given Pagination
-// parameters
-func pageNumberAndSize(offset, maxNum int) (int, int) {
-	pageNumber := 1
-	pageSize := 100
-	if offset > 0 {
-		if maxNum == -1 {
-			pageSize = offset
-			pageNumber = 2
-		} else {
-			pageSize = maxNum
-			pageNumber = offset/pageSize + 1
-		}
-	} else {
-		if maxNum > 0 {
-			pageSize = maxNum
-		}
-	}
-	return pageNumber, pageSize
-}
-
 func (db Database) SpecificSearchParameters() map[string]interface{} {
 	return map[string]interface{}{
 		// see https://files.jgi.doe.gov/apidoc/#/GET/search_list
@@ -523,103 +102,6 @@ func (db Database) SpecificSearchParameters() map[string]interface{} {
 		"s":                    []string{"name", "id", "title", "kingdom", "score.avg"}, // sort order
 		"extra":                []string{"img_taxon_oid", "project_id"},                 // list of requested extra fields
 	}
-}
-
-// checks JDP-specific search parameters and adds them to the given URL values
-func (db Database) addSpecificSearchParameters(params map[string]json.RawMessage, p *url.Values) error {
-	paramSpec := db.SpecificSearchParameters()
-	for name, jsonValue := range params {
-		switch name {
-		case "f": // field-specific search
-			var value string
-			err := json.Unmarshal(jsonValue, &value)
-			if err != nil {
-				return &databases.InvalidSearchParameter{
-					Database: "JDP",
-					Message:  "Invalid search field given (must be string)",
-				}
-			}
-			acceptedValues := paramSpec["f"].([]string)
-			if slices.Contains(acceptedValues, value) {
-				p.Add(name, value)
-			} else {
-				return &databases.InvalidSearchParameter{
-					Database: "JDP",
-					Message:  fmt.Sprintf("Invalid search field given: %s", value),
-				}
-			}
-		case "s": // sort order
-			var value string
-			err := json.Unmarshal(jsonValue, &value)
-			if err != nil {
-				return &databases.InvalidSearchParameter{
-					Database: "JDP",
-					Message:  "Invalid JDP sort order given (must be string)",
-				}
-			}
-			acceptedValues := paramSpec["s"].([]string)
-			if slices.Contains(acceptedValues, value) {
-				p.Add(name, value)
-			} else {
-				return &databases.InvalidSearchParameter{
-					Database: "JDP",
-					Message:  fmt.Sprintf("Invalid JDP sort order: %s", value),
-				}
-			}
-		case "d": // sort direction
-			var value string
-			err := json.Unmarshal(jsonValue, &value)
-			if err != nil {
-				return &databases.InvalidSearchParameter{
-					Database: "JDP",
-					Message:  "Invalid JDP sort direction given (must be string)",
-				}
-			}
-			acceptedValues := paramSpec["d"].([]string)
-			if slices.Contains(acceptedValues, value) {
-				p.Add(name, value)
-			} else {
-				return &databases.InvalidSearchParameter{
-					Database: "JDP",
-					Message:  fmt.Sprintf("Invalid JDP sort direction: %s", value),
-				}
-			}
-		case "include_private_data": // search for private data
-			var value int
-			err := json.Unmarshal(jsonValue, &value)
-			if err != nil || (value != 0 && value != 1) {
-				return &databases.InvalidSearchParameter{
-					Database: "JDP",
-					Message:  "Invalid flag given for include_private_data (must be 0 or 1)",
-				}
-			}
-			p.Add(name, fmt.Sprintf("%d", value))
-		case "extra": // comma-separated additional fields requested
-			var value string
-			err := json.Unmarshal(jsonValue, &value)
-			if err != nil {
-				return &databases.InvalidSearchParameter{
-					Database: "JDP",
-					Message:  "Invalid JDP requested extra field given (must be comma-delimited string)",
-				}
-			}
-			acceptedValues := paramSpec["extra"].([]string)
-			if slices.Contains(acceptedValues, value) {
-				p.Add(name, value)
-			} else {
-				return &databases.InvalidSearchParameter{
-					Database: "JDP",
-					Message:  fmt.Sprintf("Invalid requested extra field: %s", value),
-				}
-			}
-		default:
-			return &databases.InvalidSearchParameter{
-				Database: "JDP",
-				Message:  fmt.Sprintf("Unrecognized JDP-specific search parameter: %s", name),
-			}
-		}
-	}
-	return nil
 }
 
 func (db *Database) Search(params databases.SearchParameters) (databases.SearchResults, error) {
@@ -835,4 +317,507 @@ func (db *Database) StagingStatus(id uuid.UUID) (databases.StagingStatus, error)
 func (db *Database) LocalUser(orcid string) (string, error) {
 	// no current mechanism for this
 	return "localuser", nil
+}
+
+//--------------------
+// Internal machinery
+//--------------------
+
+const (
+	jdpBaseURL     = "https://files.jgi.doe.gov/"
+	filePathPrefix = "/global/dna/dm_archive/" // directory containing JDP files
+)
+
+// a mapping from file suffixes to format labels
+var suffixToFormat = map[string]string{
+	"bam":      "bam",
+	"bam.bai":  "bai",
+	"blasttab": "blast",
+	"bz":       "bzip",
+	"bz2":      "bzip2",
+	"csv":      "csv",
+	"faa":      "fasta",
+	"fasta":    "fasta",
+	"fasta.gz": "fasta",
+	"fastq":    "fastq",
+	"fastq.gz": "fastq",
+	"fna":      "fasta",
+	"gff":      "gff",
+	"gff3":     "gff3",
+	"gz":       "gz",
+	"html":     "html",
+	"info":     "texinfo",
+	"out":      "text",
+	"pdf":      "pdf",
+	"tar":      "tar",
+	"tar.gz":   "tar",
+	"tar.bz":   "tar",
+	"tar.bz2":  "tar",
+	"tsv":      "tsv",
+	"txt":      "text",
+}
+
+// this gets populated automatically with the keys in suffixToFormat
+var supportedSuffixes []string
+
+// a mapping from file format labels to mime types
+var formatToMimeType = map[string]string{
+	"bam":   "application/octet-stream",
+	"bai":   "application/octet-stream",
+	"csv":   "text/csv",
+	"fasta": "text/plain",
+	"fastq": "text/plain",
+	"gff":   "text/plain",
+	"gff3":  "text/plain",
+	"gz":    "application/gzip",
+	"bz":    "application/x-bzip",
+	"bz2":   "application/x-bzip2",
+	"tar":   "application/x-tar",
+	"text":  "text/plain",
+}
+
+// a mapping from the JDP's reported file types to mime types
+// (backup method for determining mime types)
+var fileTypeToMimeType = map[string]string{
+	"text":     "text/plain",
+	"fasta":    "text/plain",
+	"fasta.gz": "application/gzip",
+	"fastq":    "text/plain",
+	"fastq.gz": "application/gzip",
+	"tab":      "text/plain",
+	"tar.gz":   "application/x-tar",
+	"tar.bz":   "application/x-tar",
+	"tar.bz2":  "application/x-tar",
+}
+
+// extracts the file format from the name and type of the file
+func formatFromFileName(fileName string) string {
+	// make a list of the supported suffixes if we haven't yet
+	if supportedSuffixes == nil {
+		supportedSuffixes = make([]string, 0)
+		for suffix := range suffixToFormat {
+			supportedSuffixes = append(supportedSuffixes, suffix)
+		}
+	}
+
+	// determine whether the file matches any of the supported suffixes,
+	// selecting the longest matching suffix
+	format := "unknown"
+	longestSuffix := 0
+	for _, suffix := range supportedSuffixes {
+		if strings.HasSuffix(fileName, suffix) && len(suffix) > longestSuffix {
+			format = suffixToFormat[suffix]
+			longestSuffix = len(suffix)
+		}
+	}
+	return format
+}
+
+// extracts the file format from the name and type of the file
+func mimeTypeFromFormatAndTypes(format string, fileTypes []string) string {
+	// try to match the file type to a mime type
+	for _, fileType := range fileTypes {
+		if mimeType, ok := fileTypeToMimeType[fileType]; ok {
+			return mimeType
+		}
+	}
+	// check the file format to see whether it matches a mime type
+	if mimeType, ok := formatToMimeType[format]; ok {
+		return mimeType
+	}
+	return ""
+}
+
+// extracts file type information from the given File
+func fileTypesFromFile(_ File) []string {
+	// TODO: See https://pkg.go.dev/encoding/json?utm_source=godoc#example-RawMessage-Unmarshal
+	// TODO: for an example of how to unmarshal a variant type.
+	return []string{}
+}
+
+// extracts source information from the given metadata
+func sourcesFromMetadata(md Metadata) []frictionless.DataSource {
+	sources := make([]frictionless.DataSource, 0)
+	piInfo := md.Proposal.PI
+	if len(piInfo.LastName) > 0 {
+		var title string
+		if len(piInfo.FirstName) > 0 {
+			title = fmt.Sprintf("%s, %s", piInfo.LastName, piInfo.FirstName)
+			if len(piInfo.MiddleName) > 0 {
+				title += fmt.Sprintf(" %s", piInfo.MiddleName)
+			}
+			if len(piInfo.Institution) > 0 {
+				if len(piInfo.Country) > 0 {
+					title += fmt.Sprintf(" (%s, %s)", piInfo.Institution, piInfo.Country)
+				} else {
+					title += fmt.Sprintf(" (%s)", piInfo.Institution)
+				}
+			} else if len(piInfo.Country) > 0 {
+				title += fmt.Sprintf(" (%s)", piInfo.Country)
+			}
+		}
+		var doiURL string
+		if len(md.Proposal.AwardDOI) > 0 {
+			doiURL = fmt.Sprintf("https://doi.org/%s", md.Proposal.AwardDOI)
+		}
+		source := frictionless.DataSource{
+			Title: title,
+			Path:  doiURL,
+			Email: piInfo.EmailAddress,
+		}
+		sources = append(sources, source)
+	}
+	return sources
+}
+
+// creates a Frictionless DataResource-savvy name for a file:
+// * the name consists of lower case characters plus '.', '-', and '_'
+// * all forbidden characters encountered in the filename are removed
+// * a number suffix is added if needed to make the name unique
+func dataResourceName(filename string) string {
+	name := strings.ToLower(filename)
+
+	// remove any file suffix
+	lastDot := strings.LastIndex(name, ".")
+	if lastDot != -1 {
+		name = name[:lastDot]
+	}
+
+	// replace sequences of invalid characters with '_'
+	for {
+		isInvalid := func(c rune) bool {
+			return !unicode.IsLetter(c) && !unicode.IsDigit(c) && c != '_' && c != '-' && c != '.'
+		}
+		start := strings.IndexFunc(name, isInvalid)
+		if start >= 0 {
+			nameRunes := []rune(name)
+			end := start + 1
+			for end < len(name) && isInvalid(nameRunes[end]) {
+				end++
+			}
+			if end < len(name) {
+				name = name[:start] + string('_') + name[end+1:]
+			} else {
+				name = name[:start] + string('_')
+			}
+		} else {
+			break
+		}
+	}
+
+	return name
+}
+
+// creates a DataResource from a File
+func dataResourceFromFile(file File) frictionless.DataResource {
+	id := "JDP:" + file.Id
+	format := formatFromFileName(file.Name)
+	fileTypes := fileTypesFromFile(file)
+	sources := sourcesFromMetadata(file.Metadata)
+
+	// we use relative file paths in accordance with the Frictionless
+	// Data Resource specification
+	filePath := filepath.Join(strings.TrimPrefix(file.Path, filePathPrefix), file.Name)
+
+	pi := file.Metadata.Proposal.PI
+	return frictionless.DataResource{
+		Id:        id,
+		Name:      dataResourceName(file.Name),
+		Path:      filePath,
+		Format:    format,
+		MediaType: mimeTypeFromFormatAndTypes(format, fileTypes),
+		Bytes:     file.Size,
+		Hash:      file.MD5Sum,
+		Sources:   sources,
+		Credit: credit.CreditMetadata{
+			Identifier:   id,
+			ResourceType: "dataset",
+			Titles: []credit.Title{
+				{
+					Title: filePath,
+				},
+			},
+			Dates: []credit.EventDate{
+				{
+					Date:  file.Date,
+					Event: "Created",
+				},
+				{
+					Date:  file.AddedDate,
+					Event: "Accepted",
+				},
+				{
+					Date:  file.ModifiedDate,
+					Event: "Updated",
+				},
+			},
+			Publisher: credit.Organization{
+				OrganizationId:   "ROR:04xm1d337",
+				OrganizationName: "Joint Genome Institute",
+			},
+			RelatedIdentifiers: []credit.PermanentID{
+				{
+					Id:               file.Metadata.Proposal.DOI,
+					Description:      "Proposal DOI",
+					RelationshipType: "IsCitedBy",
+				},
+				{
+					Id:               file.Metadata.Proposal.AwardDOI,
+					Description:      "Awarded proposal DOI",
+					RelationshipType: "IsCitedBy",
+				},
+			},
+			Contributors: []credit.Contributor{
+				{
+					ContributorType: "Person",
+					// ContributorId: nothing yet
+					Name:       strings.TrimSpace(fmt.Sprintf("%s, %s %s", pi.LastName, pi.FirstName, pi.MiddleName)),
+					GivenName:  strings.TrimSpace(fmt.Sprintf("%s %s", pi.FirstName, pi.MiddleName)),
+					FamilyName: strings.TrimSpace(pi.LastName),
+					Affiliations: []credit.Organization{
+						{
+							OrganizationName: pi.Institution,
+						},
+					},
+					ContributorRoles: "PI",
+				},
+			},
+			Version: file.Date,
+		},
+	}
+}
+
+// adds an appropriate authorization header to given HTTP request
+func (db Database) addAuthHeader(request *http.Request) {
+	if len(db.Secret) > 0 { // use shared secret
+		request.Header.Add("Authorization", fmt.Sprintf("Token %s_%s", db.Orcid, db.Secret))
+	} else { // try SSO token
+		request.Header.Add("Authorization", fmt.Sprintf("Bearer %s", db.SsoToken))
+	}
+}
+
+// performs a GET request on the given resource, returning the resulting
+// response and error
+func (db *Database) get(resource string, values url.Values) (*http.Response, error) {
+	var u *url.URL
+	u, err := url.ParseRequestURI(jdpBaseURL)
+	if err != nil {
+		return nil, err
+	}
+	u.Path = resource
+	u.RawQuery = values.Encode()
+	res := fmt.Sprintf("%v", u)
+	slog.Debug(fmt.Sprintf("GET: %s", res))
+	req, err := http.NewRequest(http.MethodGet, res, http.NoBody)
+	if err != nil {
+		return nil, err
+	}
+	db.addAuthHeader(req)
+	return db.Client.Do(req)
+}
+
+// performs a POST request on the given resource, returning the resulting
+// response and error
+func (db *Database) post(resource string, body io.Reader) (*http.Response, error) {
+	u, err := url.ParseRequestURI(jdpBaseURL)
+	if err != nil {
+		return nil, err
+	}
+	u.Path = resource
+	res := fmt.Sprintf("%v", u)
+	slog.Debug(fmt.Sprintf("POST: %s", res))
+	req, err := http.NewRequest(http.MethodPost, res, body)
+	if err != nil {
+		return nil, err
+	}
+	db.addAuthHeader(req)
+	req.Header.Set("Content-Type", "application/json")
+	return db.Client.Do(req)
+}
+
+// this helper extracts files for the JDP /search GET query with given parameters
+func (db *Database) filesFromSearch(params url.Values) (databases.SearchResults, error) {
+	var results databases.SearchResults
+
+	idEncountered := make(map[string]bool) // keep track of duplicates
+
+	// extract any requested "extra" metadata fields (and scrub them from params)
+	var extraFields []string
+	if params.Has("extra") {
+		extraFields = strings.Split(params.Get("extra"), ",")
+		params.Del("extra")
+	}
+
+	resp, err := db.get("search", params)
+	if err != nil {
+		return results, err
+	}
+	defer resp.Body.Close()
+	var body []byte
+	body, err = io.ReadAll(resp.Body)
+	if err != nil {
+		return results, err
+	}
+	type JDPResults struct {
+		Organisms []struct {
+			Id    string `json:"id"`
+			Files []File `json:"files"`
+		} `json:"organisms"`
+	}
+	results.Resources = make([]frictionless.DataResource, 0)
+	var jdpResults JDPResults
+	err = json.Unmarshal(body, &jdpResults)
+	if err != nil {
+		return results, err
+	}
+	for _, org := range jdpResults.Organisms {
+		resources := make([]frictionless.DataResource, 0)
+		for _, file := range org.Files {
+			res := dataResourceFromFile(file)
+
+			// add any requested additional metadata
+			if extraFields != nil {
+				extras := "{"
+				for i, field := range extraFields {
+					if i > 0 {
+						extras += ", "
+					}
+					switch field {
+					case "project_id":
+						extras += fmt.Sprintf(`"project_id": "%s"`, org.Id)
+					case "img_taxon_oid":
+						extras += fmt.Sprintf(`"img_taxon_oid": %d`, file.Metadata.IMG.TaxonOID)
+					}
+				}
+				extras += "}"
+				res.Extra = json.RawMessage(extras)
+			}
+
+			// add the resource to our results if it's not there already
+			if _, encountered := idEncountered[res.Id]; !encountered {
+				resources = append(resources, res)
+				idEncountered[res.Id] = true
+			}
+		}
+		results.Resources = append(results.Resources, resources...)
+	}
+	return results, nil
+}
+
+// returns the page number and page size corresponding to the given Pagination
+// parameters
+func pageNumberAndSize(offset, maxNum int) (int, int) {
+	pageNumber := 1
+	pageSize := 100
+	if offset > 0 {
+		if maxNum == -1 {
+			pageSize = offset
+			pageNumber = 2
+		} else {
+			pageSize = maxNum
+			pageNumber = offset/pageSize + 1
+		}
+	} else {
+		if maxNum > 0 {
+			pageSize = maxNum
+		}
+	}
+	return pageNumber, pageSize
+}
+
+// checks JDP-specific search parameters and adds them to the given URL values
+func (db Database) addSpecificSearchParameters(params map[string]json.RawMessage, p *url.Values) error {
+	paramSpec := db.SpecificSearchParameters()
+	for name, jsonValue := range params {
+		switch name {
+		case "f": // field-specific search
+			var value string
+			err := json.Unmarshal(jsonValue, &value)
+			if err != nil {
+				return &databases.InvalidSearchParameter{
+					Database: "JDP",
+					Message:  "Invalid search field given (must be string)",
+				}
+			}
+			acceptedValues := paramSpec["f"].([]string)
+			if slices.Contains(acceptedValues, value) {
+				p.Add(name, value)
+			} else {
+				return &databases.InvalidSearchParameter{
+					Database: "JDP",
+					Message:  fmt.Sprintf("Invalid search field given: %s", value),
+				}
+			}
+		case "s": // sort order
+			var value string
+			err := json.Unmarshal(jsonValue, &value)
+			if err != nil {
+				return &databases.InvalidSearchParameter{
+					Database: "JDP",
+					Message:  "Invalid JDP sort order given (must be string)",
+				}
+			}
+			acceptedValues := paramSpec["s"].([]string)
+			if slices.Contains(acceptedValues, value) {
+				p.Add(name, value)
+			} else {
+				return &databases.InvalidSearchParameter{
+					Database: "JDP",
+					Message:  fmt.Sprintf("Invalid JDP sort order: %s", value),
+				}
+			}
+		case "d": // sort direction
+			var value string
+			err := json.Unmarshal(jsonValue, &value)
+			if err != nil {
+				return &databases.InvalidSearchParameter{
+					Database: "JDP",
+					Message:  "Invalid JDP sort direction given (must be string)",
+				}
+			}
+			acceptedValues := paramSpec["d"].([]string)
+			if slices.Contains(acceptedValues, value) {
+				p.Add(name, value)
+			} else {
+				return &databases.InvalidSearchParameter{
+					Database: "JDP",
+					Message:  fmt.Sprintf("Invalid JDP sort direction: %s", value),
+				}
+			}
+		case "include_private_data": // search for private data
+			var value int
+			err := json.Unmarshal(jsonValue, &value)
+			if err != nil || (value != 0 && value != 1) {
+				return &databases.InvalidSearchParameter{
+					Database: "JDP",
+					Message:  "Invalid flag given for include_private_data (must be 0 or 1)",
+				}
+			}
+			p.Add(name, fmt.Sprintf("%d", value))
+		case "extra": // comma-separated additional fields requested
+			var value string
+			err := json.Unmarshal(jsonValue, &value)
+			if err != nil {
+				return &databases.InvalidSearchParameter{
+					Database: "JDP",
+					Message:  "Invalid JDP requested extra field given (must be comma-delimited string)",
+				}
+			}
+			acceptedValues := paramSpec["extra"].([]string)
+			if slices.Contains(acceptedValues, value) {
+				p.Add(name, value)
+			} else {
+				return &databases.InvalidSearchParameter{
+					Database: "JDP",
+					Message:  fmt.Sprintf("Invalid requested extra field: %s", value),
+				}
+			}
+		default:
+			return &databases.InvalidSearchParameter{
+				Database: "JDP",
+				Message:  fmt.Sprintf("Unrecognized JDP-specific search parameter: %s", name),
+			}
+		}
+	}
+	return nil
 }
