+++ databases/jdp/errors.go
@@ -1,0 +1,45 @@
+// Copyright (c) 2023 The KBase Project and its Contributors
+// Copyright (c) 2023 Cohere Consulting, LLC
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy of
+// this software and associated documentation files (the "Software"), to deal in
+// the Software without restriction, including without limitation the rights to
+// use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+// of the Software, and to permit persons to whom the Software is furnished to do
+// so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+package jdp
+
+import (
+	"fmt"
+)
+
+// this error type is returned when a file is requested for which the requester
+// does not have permission
+type PermissionDeniedError struct {
+	fileId string
+}
+
+func (e PermissionDeniedError) Error() string {
+	return fmt.Sprintf("Can't access file %s: permission denied.", e.fileId)
+}
+
+// this error type is returned when a file is requested and is not found
+type FileIdNotFoundError struct {
+	fileId string
+}
+
+func (e FileIdNotFoundError) Error() string {
+	return fmt.Sprintf("Can't access file %s: not found.", e.fileId)
+}
